Hiding save button and yes & no buttons on grand prix bet screen in preview bet mode.
